git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@15419 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/app/Decorator.cpp
@@ -288,8 +288,8 @@ Decorator::GetZoom()
 
 // GetSizeLimits
 void
-Decorator::GetSizeLimits(float* minWidth, float* minHeight,
-						 float* maxWidth, float* maxHeight) const
+Decorator::GetSizeLimits(int32* minWidth, int32* minHeight,
+						 int32* maxWidth, int32* maxHeight) const
 {
 }
